Add support for unicode line separator to SmartCommitTokenizer
A string ending in a [line separator], a whitespace unicode character (everyone’s favourite kind), would result in a syntax error.

This fixes that by matching tokenizing the `workLogComment` using a “not whitespace” regex pattern. This pattern matches any character that isn’t a carriage return, newline, tab, or space. You know, the normal whitespace stuff.

This results in the line separator being included as a work log comment but because we’re currently only extracting the issue keys from the tokenized smart commit, it will have no ill effect.

[Unicode line separator]: https://www.fileformat.info/info/unicode/char/2028/index.htm